relocate hook o feature infrastructure layer
