+++ src/features/pokemon-list/infrastructure/react/hooks/__tests__/usePokemonList.isError.test.ts
@@ -1,7 +1,7 @@
 import { renderHook, waitFor } from "@testing-library/react";
 import { vi, it, expect } from "vitest";
 import usePokemonList from "../usePokemonList";
-import { PokemonRepository } from "../../../domain/ports/PokemonRepository";
+import { PokemonRepository } from "../../../../domain/ports/PokemonRepository";
 import { testData } from "./setupTests";
 
 it("starts as false when no selectedType is provided", () => {
