+++ src/features/pokemon-list/infrastructure/react/hooks/__tests__/usePokemonList.test.ts
@@ -1,10 +1,10 @@
 import { renderHook, waitFor } from "@testing-library/react";
 import { vi, it, expect } from "vitest";
 import usePokemonList from "../usePokemonList";
-import { PokemonRepository } from "../../../domain/ports/PokemonRepository";
+import { PokemonRepository } from "../../../../domain/ports/PokemonRepository";
 
-import { PokemonByType } from "../../../domain/value-objects/PokemonByType";
-import { PokemonByName } from "../../../domain/value-objects/PokemonByName";
+import { PokemonByType } from "../../../../domain/value-objects/PokemonByType";
+import { PokemonByName } from "../../../../domain/value-objects/PokemonByName";
 import { testData } from "./setupTests";
 
 it("returns a list of Pokemon items with the required values", async () => {
